+++ mrb_doc/models/image.rb
@@ -126,6 +126,14 @@ class Image
     self
   end
 
+  # Pre-multiplies the alpha for the image
+  # @return [Image]
+  def premultiply_alpha!
+    # src/mruby_integration/models/image.cpp
+    self
+  end
+
+
   # Generates a new image of width by height in the specified colour.
   # @param width [Integer]
   # @param height [Integer]